src/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Khthonian <anansicorp@tuta.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pydantic = {extras = ["email"], version = "^2.6.3"}
pydantic-extra-types = "^2.6.0"
requests = "^2.31.0"
pillow = "^10.2.0"
websockets = "^12.0"
sqlalchemy = "^2.0.27"
databases = "^0.9.0"
jinja2 = "^3.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
ipdata = "^4.0.7"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
openai = "^1.13.3"
ratelimit = "^2.2.1"
slowapi = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
